Move test resources
